refactor: CategoryTabActiveBadge를 ContentList로 이동
