+++ src/components/ContentList/SubscribedContentList.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect, cc, ce, Main } from "@/libs";
 import { MediaIdByCategories } from "@/models/Newsstand";
-import { CategoryTabActiveBadge } from "../MediaContent/CategoryTabActiveBadge";
+import { CategoryTabActiveBadge } from "./CategoryTabActiveBadge";
 import { MediaContent } from "../MediaContent/MediaContent";
 import { MediaContentTabs } from "./ContentListHeader";
 import { getSubscribedMedias } from "@/remotes/getSubscribedMedias";
